Deprecated tsc4j spring-boot-starter, moved tests in a different sub-package.
