+++ spring-boot-starter-tsc4j/src/main/resources/META-INF/spring.factories
@@ -1,9 +1,3 @@
 # Auto Configure
-#org.springframework.boot.autoconfigure.EnableAutoConfiguration=\
-#com.github.tsc4j.spring.Tsc4jSpringConfiguration
-
-# Bootstrap components
-#org.springframework.cloud.bootstrap.BootstrapConfiguration=\
-#  com.github.tsc4j.spring.Tsc4jBootstrapSpringConfiguration
-
-# EOF
+org.springframework.boot.autoconfigure.EnableAutoConfiguration=\
+    com.github.tsc4j.springboot.WarnAboutDeprecation
