+++ spring-boot-starter-tsc4j/src/test/groovy/com/github/tsc4j/springboot/ImmutableBean.java
@@ -1,0 +1,36 @@
+/*
+ * Copyright 2017 - 2021 tsc4j project
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *   http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package com.github.tsc4j.springboot;
+
+
+import com.github.tsc4j.api.Tsc4jBeanBuilder;
+import lombok.Builder;
+import lombok.Value;
+import org.springframework.boot.context.properties.ConfigurationProperties;
+
+@Value
+@Builder
+@Tsc4jBeanBuilder
+//@Component
+@ConfigurationProperties("test.toBean")
+public class ImmutableBean {
+    private boolean aBoolean;
+    private int aInt;
+    private long aLong;
+    private double aDouble;
+    private String aString;
+}
