+++ spring-boot-starter-tsc4j/src/test/groovy/com/github/tsc4j/springboot/SampleApp.groovy
@@ -1,5 +1,5 @@
 /*
- * Copyright 2017 - 2019 tsc4j project
+ * Copyright 2017 - 2021 tsc4j project
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -12,10 +12,9 @@
  * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
  * See the License for the specific language governing permissions and
  * limitations under the License.
- *
  */
 
-package com.github.tsc4j.spring
+package com.github.tsc4j.springboot
 
 import com.github.tsc4j.api.ReloadableConfig
 import com.typesafe.config.Config
