+++ spring-boot-starter-tsc4j/src/test/groovy/com/github/tsc4j/springboot/SpringAppSpec.groovy
@@ -1,5 +1,5 @@
 /*
- * Copyright 2017 - 2019 tsc4j project
+ * Copyright 2017 - 2021 tsc4j project
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -12,11 +12,14 @@
  * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
  * See the License for the specific language governing permissions and
  * limitations under the License.
- *
  */
 
-package com.github.tsc4j.spring
+package com.github.tsc4j.springboot
 
+import com.github.tsc4j.api.ReloadableConfig
+import com.github.tsc4j.spring.Tsc4jHealthIndicator
+import com.github.tsc4j.spring.Tsc4jSpringContextRefresher
+import com.typesafe.config.Config
 import groovy.util.logging.Slf4j
 import org.springframework.beans.factory.annotation.Autowired
 import org.springframework.context.ApplicationContext
@@ -31,6 +34,11 @@ class SpringAppSpec extends SpringSpec {
     def "context should be wired"() {
         expect:
         ctx != null
+
+        ctx.getBean(ReloadableConfig) != null
+        ctx.getBean(Config) != null
+        ctx.getBean(Tsc4jHealthIndicator) != null
+        ctx.getBean(Tsc4jSpringContextRefresher) != null
     }
 
     def "spring context should return correctly extract configuration property values from hocon config"() {
@@ -40,7 +48,7 @@ class SpringAppSpec extends SpringSpec {
         then:
         verifyAll {
             name == "mySuperFunkyApp"
-            ctx.getEnvironment().getProperty("server.port", Integer) == 8080
+            ctx.getEnvironment().getProperty("server.port", Integer) == -1
             ctx.getEnvironment().getProperty('app.var2', Integer) == 42
             ctx.getEnvironment().getProperty("app.var3") == "overriden in funky/application.conf: mySuperFunkyApp"
         }
