+++ spring-boot-starter-tsc4j/src/test/groovy/com/github/tsc4j/springboot/SpringSpec.groovy
@@ -1,5 +1,5 @@
 /*
- * Copyright 2017 - 2019 tsc4j project
+ * Copyright 2017 - 2021 tsc4j project
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -12,22 +12,15 @@
  * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
  * See the License for the specific language governing permissions and
  * limitations under the License.
- *
  */
 
-package com.github.tsc4j.spring;
+package com.github.tsc4j.springboot
 
-import lombok.Data;
-import org.springframework.boot.context.properties.ConfigurationProperties;
-import org.springframework.stereotype.Component;
+import org.springframework.boot.test.context.SpringBootTest
+import org.springframework.test.context.ActiveProfiles
+import spock.lang.Specification
 
-@Data
-@Component
-@ConfigurationProperties("test.bean")
-public class StandardBean {
-    private boolean aBoolean;
-    private int aInt;
-    private long aLong;
-    private double aDouble;
-    private String aString;
+@ActiveProfiles(["test", "funky"])
+@SpringBootTest(classes = SampleApp)
+abstract class SpringSpec extends Specification {
 }
